Move app config to its own package
